+++ src/Suru.Compiler/Lex/TokenKind.cs
@@ -1,0 +1,6 @@
+namespace Suru.Compiler.Lex;
+
+public enum TokenKind
+{
+    Eof,
+}